lambda/requirements.txt
yfinance==0.2.36
psycopg2-binary==2.9.9
pandas==2.2.1